axf/requirement.txt
cffi==1.14.0
cryptography==2.8
Django==1.11.7
django-debug-toolbar==2.2
django-redis==4.11.0
django-redis-cache==2.1.0
django-tinymce==2.8.0
Pillow==7.0.0
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
uWSGI==2.0.18
